% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignmnent_feedback.R
\name{get_assignmnent_feedback}
\alias{get_assignmnent_feedback}
\title{Call the Gradio backend}
\usage{
get_assignmnent_feedback(homework_name, student_answer, container_name)
}
\arguments{
\item{homework_name}{Homework name of format homeworkXqY
where X is the homework number and Y is the question number.}

\item{student_answer}{Student answer.}

\item{container_name}{Name of container.}
}
\description{
Function to call the Gradio backend LLM with the
homework_name and student answer.
}
\details{
The feedback returned is in a funky format because
Gradio's API only supports streaming results so we
don't get nice clean JSON. Instead we get a line of
text with the event status followed by the actual
LLM with \verb{\\n} and \verb{"} escaped something like the following:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{"event: complete\\ndata: [\\"The question is:\\\\n\\\\nThe `midwest` data frame...\\n]\\n\\n"
}\if{html}{\out{</div>}}
}
\examples{
feedback <- get_assignmnent_feedback(
  homework_name = "homework1-q1",
  student_answer = "give me the question",
  container_name = "container_hostname_as_a_session_id"
)
feedback
}
