% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feedback.R
\name{get_feedback}
\alias{get_feedback}
\title{Capture selected text and launch Shiny app for feedback}
\usage{
get_feedback()
}
\value{
A character vector containing the selected text. If no text is
selected, returns an empty character vector.
}
\description{
An RStudio addin that captures the currently selected text in the editor
or console, returns it as a character vector that is stored as a
system variable, that is then passed on to the feedback getting
Shiny app that is launched.
}
\details{
Users interact with this function through the Get feedback addin.
}
