% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hostname.R
\name{get_hostname}
\alias{get_hostname}
\title{Get the container hostname we are running inside}
\usage{
get_hostname()
}
\value{
nodename
}
\description{
Get the container hostname we are running inside
}
